+++ docs/make.jl
@@ -1,5 +1,3 @@
-push!(LOAD_PATH, "../src/")
-
 using Documenter
 using Titan
 
@@ -7,15 +5,4 @@ makedocs(
     sitename = "Titan.jl",
     authors = "Damir Akchurin",
     pages = [
-        "Home" => "index.md",
-        "Quick Start" => "QuickStart.md",
-        "Documentation" => [
-            "Model" => "Model.md",
-            "Materials" => "Materials.md",
-            "Sections" => "Sections.md",
-            "Elements" => "Elements.md",
-            "Loads" => "Loads.md"],
-        "Automatic Differentiation" => "AutomaticDifferentiation.md"])
-
-deploydocs(
-    repo="github.com/akchurinda/Titan.jl")
+        "Home" => "index.md"])