+++ src/Analysis/Analysis.jl
@@ -1,0 +1,6 @@
+abstract type AbstractAnalysis end
+
+include("./Assembling.jl")
+include("./LinearElasticAnalysis.jl")
+include("./NonlinearElasticAnalysis.jl")
+include("./ElasticBucklingAnalysis.jl")