+++ src/Components/Nodes.jl
@@ -1,0 +1,7 @@
+struct Node
+    x::Real
+    y::Real
+    state::NodeState
+
+    Node(x::Real, y::Real) = new(x, y, NodeState())
+end