+++ src/Components/Sections.jl
@@ -1,0 +1,4 @@
+struct Section
+    A::Real
+    I::Real
+end