+++ src/Materials.jl
@@ -1,4 +1,0 @@
-abstract type AbstractMaterial{T <: Real} end
-
-include("Material Library/ElasticMaterial.jl")
-include("Material Library/ElasticPerfectlyPlasticMaterial.jl")