+++ NEWS.md
@@ -1,1 +1,6 @@
 # V0.1.0
+
+- Better handing of components.
+- Working version of linear elastic analysis that is automatically differentiable using `ForwardDiff.jl` and `ReverseDiff.jl`.
+- Working version of nonlinear elastic analysis that is automatically differentiable using `ForwardDiff.jl` and `ReverseDiff.jl`.
+- Plotting extension of undeformed and deformed shapes enabled using `Makie.jl` extension.